fix for projects, organizations segments table

+++ backend/src/database/repositories/memberRepository.ts
@@ -26,7 +26,6 @@ import {
 } from './types/memberTypes'
 import { ActivityDisplayVariant } from '../../types/activityTypes'
 import SegmentRepository from './segmentRepository'
-import { SegmentData } from '../../types/segmentTypes'
 
 const { Op } = Sequelize
 
@@ -160,6 +159,25 @@ class MemberRepository {
       type: QueryTypes.INSERT,
       transaction,
     })
+  }
+
+  static async excludeMemberFromSegments(memberId: string, options: IRepositoryOptions) {
+    const seq = SequelizeRepository.getSequelize(options)
+
+    const transaction = SequelizeRepository.getTransaction(options)
+
+    const bulkDeleteMemberSegments = `DELETE FROM "memberSegments" WHERE "memberId" = :memberId and "segmentId" in (:segmentIds);`
+
+    await seq.query(bulkDeleteMemberSegments, {
+      replacements: {
+        memberId,
+        segmentIds: options.currentSegments.map((s) => s.id),
+      },
+      type: QueryTypes.DELETE,
+      transaction,
+    })
+
+    return this.findById(memberId, options, true, false)
   }
 
   static async findSampleDataMemberIds(options: IRepositoryOptions) {
@@ -432,7 +450,13 @@ class MemberRepository {
     const currentTenant = SequelizeRepository.getCurrentTenant(options)
 
     const query = `
-    with identities as (select mi."memberId",
+    with segment_ids as (
+      select "memberId", array_agg("segmentId") as "segmentIds" from
+      "memberSegments"
+      where "tenantId" = :tenantId
+      group by "memberId"
+    ),
+    identities as (select mi."memberId",
                            array_agg(distinct mi.platform)             as identities,
                            jsonb_object_agg(mi.platform, mi.usernames) as username
                     from (select "memberId",
@@ -464,10 +488,12 @@ class MemberRepository {
             m."tenantId",
             m."createdById",
             m."updatedById",
-            i.username
+            i.username,
+            si."segmentIds" as segments
       from members m
               inner join "memberIdentities" mi on m.id = mi."memberId"
               inner join identities i on i."memberId" = m.id
+              inner join segment_ids si on si."memberId" = m.id
       where mi."tenantId" = :tenantId
         and mi.platform = :platform
         and mi.username in (:usernames)
@@ -515,7 +541,6 @@ class MemberRepository {
       where: {
         id,
         tenantId: currentTenant.id,
-        segmentId: options.currentSegments.map((s) => s.id),
       },
       transaction,
     })
@@ -587,6 +612,11 @@ class MemberRepository {
         transaction,
       })
     }
+
+    if (data.segments) {
+      await MemberRepository.includeMemberToSegments(record.id, { ...options, transaction })
+    }
+
     const seq = SequelizeRepository.getSequelize(options)
 
     if (data.username) {
@@ -664,25 +694,30 @@ class MemberRepository {
 
     const currentTenant = SequelizeRepository.getCurrentTenant(options)
 
-    const record = await options.database.member.findOne({
-      where: {
-        id,
-        tenantId: currentTenant.id,
-        segmentId: options.currentSegments.map((s) => s.id),
-      },
-      transaction,
-    })
+    const member = await MemberRepository.excludeMemberFromSegments(id, { ...options, transaction })
 
-    if (!record) {
-      throw new Error404()
+    // if member doesn't belong to any other segment anymore, remove it
+
+    if (member.segments.length === 0) {
+      const record = await options.database.member.findOne({
+        where: {
+          id,
+          tenantId: currentTenant.id,
+          segmentId: options.currentSegments.map((s) => s.id),
+        },
+        transaction,
+      })
+
+      if (!record) {
+        throw new Error404()
+      }
+
+      await record.destroy({
+        force,
+        transaction,
+      })
+      await this._createAuditLog(AuditLogRepository.DELETE, record, record, options)
     }
-
-    await record.destroy({
-      force,
-      transaction,
-    })
-
-    await this._createAuditLog(AuditLogRepository.DELETE, record, record, options)
   }
 
   static async destroyBulk(ids, options: IRepositoryOptions, force = false) {
@@ -701,6 +736,7 @@ class MemberRepository {
     })
   }
 
+  /*
   static async getSegments(
     memberIds: string[],
     options: IRepositoryOptions,
@@ -714,7 +750,7 @@ class MemberRepository {
       select "memberId", "segmentId", "tenantId", "createdAt" from "memberSegments" where "memberId" in (:memberIds) order by "createdAt" asc;
     `
 
-    await seq.query(query, {
+    const data = await seq.query(query, {
       replacements: {
         memberIds,
       },
@@ -722,10 +758,29 @@ class MemberRepository {
       transaction,
     })
 
+    for (const id of memberIds) {
+      results.set(id, [])
+    }
+
+    for (const res of data as any[]) {
+      const { memberId, segmentId, username, sourceId, integrationId, createdAt } = res
+      const identities = results.get(memberId)
+
+      identities.push({
+        platform,
+        username,
+        sourceId,
+        integrationId,
+        createdAt,
+      })
+    }
+
+
     // TODO: complete
 
     return results
   }
+  */
 
   static async getIdentities(
     memberIds: string[],
@@ -778,7 +833,15 @@ class MemberRepository {
   ) {
     const transaction = SequelizeRepository.getTransaction(options)
 
-    const include = []
+    const include = [
+      {
+        model: options.database.segment,
+        as: 'segments',
+        through: {
+          attributes: [],
+        },
+      },
+    ]
 
     const where: any = {
       id,
@@ -787,7 +850,6 @@ class MemberRepository {
     if (!ignoreTenant) {
       const currentTenant = SequelizeRepository.getCurrentTenant(options)
       where.tenantId = currentTenant.id
-      where.segmentId = options.currentSegments.map((s) => s.id)
     }
 
     const record = await options.database.member.findOne({
